backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
requests==2.31.0
python-dotenv==1.0.0

# Security & Crypto
bcrypt==4.1.2
cryptography==42.0.5
PyJWT==2.8.0

# Blockchain & Web3
web3==6.15.1
eth-account==0.10.0
eth-keys==0.4.0

# Rate Limiting & Caching
Flask-Limiter==4.0.0
redis==5.0.1

# Missing dependencies
bleach==6.1.0

# Email Support
Flask-Mail==0.10.0

# QR Code Generation
qrcode==7.4.2
Pillow==10.2.0

# API & HTTP
httpx==0.26.0
websockets==12.0

# Excel Export
XlsxWriter==3.2.0
